jobflow_mvp/pyproject.toml
[project]
name = "jobflow-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-adk>=1.19.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.7",
]
